Convert index to PHP / HTML5
The current schema is a tangled web of nested AJAX calls and PHP. In
order to best refactor this, the main index file ought to be in PHP.

For one, this will allow us to open up an OAuth connection from the
start, rather than re-opening the connection with each AJAX call.

Additionally, the DOCTYPE has been upgraded from XHTML 1.0 Strict to
HTML5. This lends more freedom to use new tags and updated syntax.