Merge twitteroath into root directory
The simplest means of updating is to merge with a pre-existing code
base. I.e, adapting to the twitteroath library will be easier than the
reverse.

Given the previous decision to convert index to PHP, it makes sense to
move the current HTML to twitteroath's html.inc file and use the
default twitteroath index.php. This way we will a) save server requests
by only making a single connection at the session_start, and b) create
more maintainable code through separate, functional pieces.