Reorganized basic layout
Sample code has been removed and replaced with Tweetoes-specific HTML.

The "clear sessions" notice may be unecessary. If so, it ought to be
removed in a future commit. (Note: CSS has been added for <small>)

In terms of CSS, code ought to be converted to SASS or some other form
of pre-processor for better maintenance.