Return tweets from search
Updated index.php to return searched tweets based on an array of
parameters. Note: need to confirm the effect of "count" and whether it
limits the number of returned tweets.

Once successfully retrieving and displaying tweets in an <ul>, AJAX for
Twitter can be re-incorporated.

Initial fix should be as simple as updating the twitter.php include
file: $connection->get…

However, a more elegant solution would still be best. The overall code
structure needs improvement.