Remove unecessary OAuth connect files
Because this is a single-user application and all the necessary keys
are predefined, there's no need for additional user authentication.
I.e, connect.php and redirect.php (used to reset keys, etc. and open a
new connection) are unecessary.

As such, the files have been removed to a) simplify the connection
process and b) clean up the directory.

Note: Look into function of callback.php and whether or not it is
necessary to keep. test.php may be fit to remove as well.