Update Twitter AJAX error logging
The current console.log for the Twitter AJAX call combines string and
object data into a single statement. This statement does not show the
full object in the console.

In order to view the full data object, the console.log statement has
been split in two. This way it will show one line, "Error (Twitter)",
and in the following line full object notation.