Move TwitterOAuth from index.php to twitter.php
After testing, AJAX requests require that the TwitterOAuth connection
be re-opened for each request. As such, sessions are no longer required
for storing the connection and have been removed. Additionally, OAuth
is not required at the outset of the application (i.e, in index.php)
and has therefore also been removed.

OAuth is only required during AJAX requests (i.e, in twitter.php).
Thus, the original code from index.php has been adapted to twitter.php.

Parsing methods must now be updated to account for changes in the tweet
object.

Note: This update is only being made for the sake of *first* creating a
functional application and *then* optimizing afterwards. In regards to
optimization, re-establishing the OAuth connection for each AJAX
request is inefficient and ought to be simplified to the use of a
single, stored connection. The goal is simplicity (for the sake of
maintainability) and speed (for the sake of UX).