Complete update to Twitter API 1.1
In a previous commit, OAuth was said to be removed from index.php but
was not actually. index.php now only consists of an include for
html.inc.

Test code has also been removed from html.inc. HTML ough to be further
restructured for simplicity (in light of both JS and CSS).

Application is now successfully updated to Twitter API 1.1.